rm old server.rs
